refac: ketcher-templates components
